refactor: remove unused native methods
Remove unused GetWindowText and GetWindowBoundingRectangle methods.
