+++ src/MahApps.Metro/Native/WinApiHelper.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
@@ -15,7 +14,7 @@ using Windows.Win32.Foundation;
 using Windows.Win32.Graphics.Gdi;
 using Windows.Win32.UI.WindowsAndMessaging;
 
-namespace MahApps.Metro.Controls
+namespace MahApps.Metro.Native
 {
     internal static class WinApiHelper
     {
@@ -131,67 +130,6 @@ namespace MahApps.Metro.Controls
             }
 
             return (int)fallback;
-        }
-
-        /// <summary> Gets the text associated with the given window handle. </summary>
-        /// <param name="window"> The window to act on. </param>
-        /// <returns> The window text. </returns>
-        internal static string? GetWindowText(this Window? window)
-        {
-            if (window is null == false
-                && PresentationSource.FromVisual(window) is HwndSource source
-                && source.IsDisposed == false
-                && source.RootVisual is null == false
-                && source.Handle != IntPtr.Zero)
-            {
-                int bufferSize = PInvoke.GetWindowTextLength(new HWND(source.Handle)) + 1;
-                unsafe
-                {
-                    fixed (char* windowNameChars = new char[bufferSize])
-                    {
-                        PInvoke.GetWindowText(new HWND(source.Handle), windowNameChars, bufferSize);
-                        return new string(windowNameChars);
-                    }
-                }
-            }
-
-            return default;
-        }
-
-        /// <summary> Gets the text associated with the given window handle. </summary>
-        /// <param name="window"> The window to act on. </param>
-        /// <returns> The window text. </returns>
-        internal static Rect GetWindowBoundingRectangle(this Window? window)
-        {
-            var bounds = new Rect(0, 0, 0, 0);
-
-            if (window is null == false
-                && PresentationSource.FromVisual(window) is HwndSource source
-                && source.IsDisposed == false
-                && source.RootVisual is null == false
-                && source.Handle != IntPtr.Zero)
-            {
-                var rc = new RECT();
-
-                try
-                {
-                    unsafe
-                    {
-                        PInvoke.GetWindowRect((HWND)source.Handle, &rc);
-                    }
-                }
-                // Allow empty catch statements.
-#pragma warning disable 56502
-                catch (Win32Exception)
-                {
-                }
-                // Disallow empty catch statements.
-#pragma warning restore 56502
-
-                bounds = new Rect(rc.left, rc.top, rc.GetWidth(), rc.GetHeight());
-            }
-
-            return bounds;
         }
     }
 }
